# Conflicts:
#	src/main/java/monopoly/field/Estate.java
#	src/main/java/monopoly/field/Field.java
#	src/main/java/monopoly/field/Property.java

+++ src/main/java/monopoly/field/Property.java
@@ -3,9 +3,9 @@ package monopoly.field;
 import monopoly.Player;
 
 public abstract class Property extends Field {
-    protected float price;
+    protected double price;
     protected Player owner;
-    public abstract float getRent();
+    public abstract double getRent();
 
     public Property(String name){
         super(name);
